MCTS refactor pt. 3
MCTS Refactor tentatively complete.

MCTS now used a stateless node design: each node stores only the action taken and; all game logic is localized to getGame() which uses the sequence of actions from root to reconstruct the game state on demand through a replay script for both players.

no game objects are copied under this design. the root game object is reused and restored (with a saved gamestate object) over and over again.

The goal of this design is to trade cpu work for memory efficiency to allow for high levels of parallelization. so far profiler shows about 1.5x decrease in evals/sec but 1.5x increase memory efficiency.

this system is also much more flexible and scalable and leaves the door open for dynamic checkpoint nodes, and rollout driven hidden info randomization down the line.